chore: rename mapping-specific test directories
