+++ test/integrations/sources/shopify/pixelMappingTests/CheckoutEventsTests.ts
