+++ test/integrations/sources/shopify/pixelMappingTests/CheckoutStepsTests.ts
